chore(report): fix report progress bar; download and delete requests

+++ app/src/lib/components/data/sample/report/ReportAuditTable.svelte
@@ -4,6 +4,7 @@
 	import { CerebroApi, type ApiResponse } from "$lib/utils/api";
 	import { getDateTimeString, getDateTimeStringUtc } from "$lib/utils/helpers";
 	import type { Cerebro, ReportEntry } from "$lib/utils/types";
+    import { Role } from "$lib/utils/types";
 	import { getToastStore, type PopupSettings, type ToastSettings } from "@skeletonlabs/skeleton";
     import { popup } from '@skeletonlabs/skeleton';
 
@@ -119,7 +120,6 @@
         placement: 'left'
     };
 
-
 </script>
 
 <div class="table-container text- pb-4">
@@ -166,13 +166,15 @@
                                     </svg>
                                 </div>
                             </button>
-                            <button class="btn variant-outline-tertiary text-xs p-1 ml-2  [&>*]:pointer-events-none" type="button" id="deletePopup" on:click={() => deleteReportCopy(reportEntry.id)} use:popup={deletePopup}>
-                                <div class="h-3 w-3">
-                                    <svg aria-hidden="true" fill="none" stroke="currentColor" stroke-width="1.5" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-                                        <path d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" stroke-linecap="round" stroke-linejoin="round"></path>
-                                    </svg>                               
-                                </div>
-                            </button>
+                            {#if $page.data.userData.roles.includes(Role.Data)}
+                                <button class="btn variant-outline-tertiary text-xs p-1 ml-2  [&>*]:pointer-events-none" type="button" id="deletePopup" on:click={() => deleteReportCopy(reportEntry.id)} use:popup={deletePopup}>
+                                    <div class="h-3 w-3">
+                                        <svg aria-hidden="true" fill="none" stroke="currentColor" stroke-width="1.5" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                                            <path d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0" stroke-linecap="round" stroke-linejoin="round"></path>
+                                        </svg>                               
+                                    </div>
+                                </button>
+                            {/if}
                         </td>
                     </tr>
